refactor: isolate recorder project persistence
